+++ src/common/getEmptyReport/index.js
@@ -186,6 +186,11 @@ export function getEmptyReport() {
 			unique: 0,
 			usage: {},
 		},
+		widths: {
+			total: 0,
+			unique: 0,
+			usage: {},
+		},
 		letterSpacings: {
 			total: 0,
 			unique: 0,