+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `selectors.attributesUsage` metric
 - Added `attributesUsage` flag
 - Added `imports` metric
+- Added `widths` metric
+- Added `widths` flag
 
 ### Changed